+++ app/src/main/java/xhsun/gw2app/steve/backend/database/storage/StorageWrapper.java
@@ -2,7 +2,7 @@ package xhsun.gw2app.steve.backend.database.storage;
 
 import java.util.List;
 
-import xhsun.gw2api.guildwars2.err.GuildWars2Exception;
+import me.xhsun.guildwars2wrapper.error.GuildWars2Exception;
 import xhsun.gw2app.steve.backend.data.AbstractData;
 import xhsun.gw2app.steve.backend.data.AccountData;
 import xhsun.gw2app.steve.backend.data.vault.item.Countable;
@@ -72,7 +72,7 @@ public abstract class StorageWrapper<I extends AbstractData, S extends VaultItem
 	void updateRecord(List<Countable> known, List<Countable> seen, Countable info) {
 		long[] result = updateCountableRecord(known, seen, info);
 		info.setCount(result[2]);
-		info.setId(result[3]);
+		info.setId((int) result[3]);
 		if (info.getCount() > 0 && result[1] == TRUE)
 			//noinspection unchecked
 			updateDatabase((S) info, result[0] == TRUE);
